finish graph

+++ app/src/main/java/com/santos/jukebox/client/data/TopMusicsResponse.kt
@@ -8,5 +8,5 @@ data class TopMusicsResponse(
 
 data class GraphMusic(
     val name: String,
-    val percentage: Int
+    val percentage: Float
 )
